# doc/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Search Platform -- Agentic RAG API",
    "description": "REST API for an agentic Retrieval-Augmented Generation backend: Docling document preprocessing, LlamaIndex + PGVector indexing and retrieval, a CrewAI multi-agent contextual RAG pipeline (query analysis -> retrieval -> synthesis -> verification), Arize Phoenix tracing, and RAGAs-based evaluation. Also exposes an OpenAI-compatible `/v1/chat/completions` surface so OpenWebUI (or any OpenAI SDK client) can use this backend directly.",
    "version": "0.1.0",
    "contact": {
      "name": "Document Search Platform"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local docker-compose deployment"
    }
  ],
  "paths": {
...
    "/api/v1/query": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Query",
        "description": "Native rich RAG query: runs the CrewAI agentic crew and returns the answer, citations, verifier approval status, and (optionally) the full execution trace.",
        "operationId": "post_api_v1_query",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RagQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagQueryRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "post": {
        "tags": [
          "rag"
        ],
        "summary": "Search",
        "description": "Raw retrieval only (no generation) -- useful for debugging what the index contains.",
        "operationId": "post_api_v1_search",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        }
      }
    },
...
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "openwebui-compat"
        ],
        "summary": "Chat Completions",
        "description": "OpenAI-compatible surface consumed by OpenWebUI (or any OpenAI SDK client). `stream=true` is rejected with a 400 -- this endpoint returns one final, verified answer per turn.",
        "operationId": "post_v1_chat_completions",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatCompletionChoice": {
        "properties": {
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "message": {
            "$ref": "#/components/schemas/ChatMessage"
          },
          "finish_reason": {
            "default": "stop",
            "title": "Finish Reason",
            "type": "string"
          }
        },
        "required": [
          "index",
          "message"
        ],
        "title": "ChatCompletionChoice",
        "type": "object"
      },
      "ChatCompletionRequest": {
        "properties": {
          "model": {
            "default": "agentic-rag",
            "title": "Model",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 0.1,
            "title": "Temperature"
          },
          "stream": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Stream"
          }
        },
        "required": [
          "messages"
        ],
        "title": "ChatCompletionRequest",
        "type": "object"
      },
      "ChatCompletionResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "object": {
            "default": "chat.completion",
            "title": "Object",
            "type": "string"
          },
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/ChatCompletionChoice"
            },
            "title": "Choices",
            "type": "array"
          },
          "usage": {
            "$ref": "#/components/schemas/ChatCompletionUsage"
          }
        },
        "required": [
          "model",
          "choices"
        ],
        "title": "ChatCompletionResponse",
        "type": "object"
      },
      "ChatCompletionUsage": {
        "properties": {
          "prompt_tokens": {
            "default": 0,
            "title": "Prompt Tokens",
            "type": "integer"
          },
          "completion_tokens": {
            "default": 0,
            "title": "Completion Tokens",
            "type": "integer"
          },
          "total_tokens": {
            "default": 0,
            "title": "Total Tokens",
            "type": "integer"
          }
        },
        "title": "ChatCompletionUsage",
        "type": "object"
      },
      "ChatMessage": {
        "properties": {
          "role": {
            "enum": [
              "system",
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "RagQueryRequest": {
        "properties": {
          "question": {
            "title": "Question",
            "type": "string"
          },
          "chat_history": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "",
            "description": "Freeform prior-turn context, if any.",
            "title": "Chat History"
          },
          "include_trace": {
            "default": false,
            "description": "Include the full multi-agent execution trace.",
            "title": "Include Trace",
            "type": "boolean"
          }
        },
        "required": [
          "question"
        ],
        "title": "RagQueryRequest",
        "type": "object"
      },
      "RagQueryResponse": {
        "properties": {
          "question": {
            "title": "Question",
            "type": "string"
          },
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "approved": {
            "description": "Whether the Verifier agent approved the answer as grounded.",
            "title": "Approved",
            "type": "boolean"
          },
          "iterations": {
            "title": "Iterations",
            "type": "integer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRef"
            },
            "title": "Sources",
            "type": "array"
          },
          "trace": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Trace"
          }
        },
        "required": [
          "question",
          "answer",
          "approved",
          "iterations",
          "sources"
        ],
        "title": "RagQueryResponse",
        "type": "object"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "title": "Query",
            "type": "string"
          },
          "top_k": {
            "default": 6,
            "maximum": 50,
            "minimum": 1,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
      "SearchResponse": {
        "properties": {
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "results"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "SearchResultItem": {
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source_file": {
            "title": "Source File",
            "type": "string"
          },
          "section_path": {
            "title": "Section Path",
            "type": "string"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          }
        },
        "required": [
          "text",
          "score",
          "source_file",
          "section_path",
          "node_id"
        ],
        "title": "SearchResultItem",
        "type": "object"
      },
      "SourceRef": {
        "properties": {
          "source_file": {
            "title": "Source File",
            "type": "string"
          },
          "section_path": {
            "title": "Section Path",
            "type": "string"
          }
        },
        "required": [
          "source_file",
          "section_path"
        ],
        "title": "SourceRef",
        "type": "object"
      },
      "ValidationError": {
        "title": "ValidationError",
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ]
      }
    }
  }
}